Network Player + Global Player
set Global player (self), add Setting screen to change the global player. start game with playerlist
