+++ core/src/com/mygdx/game/GameBoard.java
@@ -12,9 +12,9 @@ import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.DelayedRemovalArray;
 import com.esotericsoftware.kryonet.Connection;
 import com.esotericsoftware.kryonet.Listener;
-import com.mygdx.game.actors.AddTileActor;
-import com.mygdx.game.actors.PlayerStatusActor;
-import com.mygdx.game.actors.TileActor;
+import com.mygdx.game.actor.AddTileActor;
+import com.mygdx.game.actor.PlayerStatusActor;
+import com.mygdx.game.actor.TileActor;
 import com.mygdx.game.meeple.Meeple;
 import com.mygdx.game.meeple.MeeplePlacement;
 import com.mygdx.game.network.GameClient;
