+++ core/src/com/mygdx/game/actor/TileActor.java
@@ -1,4 +1,4 @@
-package com.mygdx.game.actors;
+package com.mygdx.game.actor;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
