+++ core/src/com/mygdx/game/network/GameClient.java
@@ -23,7 +23,7 @@ public class GameClient extends AbstractGameManager{
             // 192.168.1.2 home
             //143.205.187.139 uni
             l.add(InetAddress.getByName("143.205.187.139"));
-            return l;
+            //return l;
         } catch (Exception e) {
             e.printStackTrace();
         }
