+++ core/src/com/mygdx/game/screen/ChosenMeeplePlacementScreen.java
@@ -10,14 +10,15 @@ import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
-import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 import com.mygdx.game.Carcassonne;
-import com.mygdx.game.Feature;
+import com.mygdx.game.tile.Feature;
 import com.mygdx.game.GameBoard;
-import com.mygdx.game.TileActor;
+import com.mygdx.game.meeple.MeeplePlacement;
+import com.mygdx.game.tile.Side;
+import com.mygdx.game.actor.TileActor;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -25,8 +26,6 @@ import java.util.List;
 public class ChosenMeeplePlacementScreen extends Actor implements Screen{
 
     private Stage stage;
-    private Skin skin;
-    private Stage stageUi;
     GameBoard gb;
     TileActor currentTile;
     private List<Feature> features;
@@ -34,24 +33,22 @@ public class ChosenMeeplePlacementScreen extends Actor implements Screen{
     private List<TextButton> meepleButtons;
     private Game game;
     private Screen previousScreen;
+    MeeplePlacement mp;
 
     public  ChosenMeeplePlacementScreen(Screen previousScreen, Game game, GameBoard gb) {
         this.gb = gb;
         this.previousScreen = previousScreen;
         this.game = game;
-        stage = new Stage(new ScreenViewport());
-        stageUi = new Stage(new ScreenViewport());
         Gdx.app.setLogLevel(Application.LOG_DEBUG);
-        skin = new Skin(Gdx.files.internal("skin/glassy-ui.json"));
-        currentTile = gb.getCurrentTile();
+        stage = new Stage(new ScreenViewport());
+        currentTile = gb.getPreviousTile(); //current Tile for Placing Meeple
         features = currentTile.getFeatures();
-        meepleButtons = new ArrayList<TextButton>();
+        meepleButtons = new ArrayList<>();
+        mp = new MeeplePlacement(gb);
 
-
-
-        output = new Label("What do you want to do?", Carcassonne.skin);
+        output = new Label("Where do you want to place your Meeple?", Carcassonne.skin);
         output.setAlignment(Align.center);
-        output.setY(Gdx.graphics.getHeight()/8*7);
+        output.setY(Gdx.graphics.getHeight()/8f*7f);
         output.setWidth(Gdx.graphics.getWidth());
         output.setFontScale(3);
         stage.addActor(output);
@@ -64,7 +61,7 @@ public class ChosenMeeplePlacementScreen extends Actor implements Screen{
         setMeepleTextButtons();
 
         TextButton back = new TextButton("Back", Carcassonne.skin);
-        back.setWidth(Gdx.graphics.getWidth() / 5 - 40);
+        back.setWidth(Gdx.graphics.getWidth() / 5f - 40f);
         back.setPosition(Gdx.graphics.getWidth() - back.getWidth() - 20, 40);
         back.addListener(new InputListener() {
             @Override
@@ -78,16 +75,13 @@ public class ChosenMeeplePlacementScreen extends Actor implements Screen{
             }
         });
         stage.addActor(back);
-        //Gdx.input.setInputProcessor(stage);
-
-    }
+     }
 
     public TextButton createMeeplePlacementButton(Feature feature)
     {
-        TextButton placeMeepleButton = new TextButton("Place Meeple on " + feature.getClass().getSimpleName(), Carcassonne.skin);
-
-        placeMeepleButton.setWidth(Gdx.graphics.getWidth() / 8);
-        placeMeepleButton.setHeight(Gdx.graphics.getHeight() / 8);
+        TextButton placeMeepleButton = new TextButton("On " + feature.getClass().getSimpleName() + " " +feature.getSides(), Carcassonne.skin);
+        placeMeepleButton.setWidth(Gdx.graphics.getWidth() / 8f);
+        placeMeepleButton.setHeight(Gdx.graphics.getHeight() / 8f);
 
         placeMeepleButton.addListener(new InputListener() {
             @Override
@@ -97,10 +91,13 @@ public class ChosenMeeplePlacementScreen extends Actor implements Screen{
 
             @Override
             public void touchUp(InputEvent event, float x, float y, int pointer, int button) {
-                if (gb.getCurrentPlayer().checkIfMeepleAvailable() == true) {
-                    feature.addMeeple();
-
-                }
+                ArrayList <Side> sides;
+                sides = feature.getSides();
+                Side side = sides.get(0);
+                currentTile = gb.getPreviousTile();
+                side = currentTile.getSideAfterRotation(side);
+                mp.placeMeeple(side, feature, gb.getCurrentTile().getPosition());
+                game.setScreen(previousScreen);
             }
         });
 
@@ -110,15 +107,23 @@ public class ChosenMeeplePlacementScreen extends Actor implements Screen{
     private void setMeepleTextButtons() {
         for (int i = 0; i < meepleButtons.size() && i < 6; i++) {
             TextButton button = meepleButtons.get(i);
-            button.setWidth(Gdx.graphics.getWidth() / 2 - 40);
-            button.setHeight(Gdx.graphics.getHeight()/5 - 40);
+            button.setWidth(Gdx.graphics.getWidth() / 2f - 40f);
+            button.setHeight(Gdx.graphics.getHeight()/5f - 40f);
             if(i%2 == 0){
-                button.setPosition(20, Gdx.graphics.getHeight() - Gdx.graphics.getHeight() /5 * (i/2+2) + 40);
+                button.setPosition(20, Gdx.graphics.getHeight() - Gdx.graphics.getHeight() /5f * (i/2+2) + 40f);
             }else{
-                button.setPosition(20 + Gdx.graphics.getWidth() / 2, Gdx.graphics.getHeight() - Gdx.graphics.getHeight() /5 * (i/2+2) + 40);
+                button.setPosition(20f + Gdx.graphics.getWidth() / 2, Gdx.graphics.getHeight() - Gdx.graphics.getHeight() /5f * (i/2+2) + 40f);
             }
             stage.addActor(button);
         }
+    }
+
+    public TileActor getCurrentTile() {
+        return currentTile;
+    }
+
+    public void setCurrentTile(TileActor currentTile) {
+        this.currentTile = currentTile;
     }
 
     @Override
@@ -130,7 +135,6 @@ public class ChosenMeeplePlacementScreen extends Actor implements Screen{
     public void render(float delta) {
         Gdx.gl.glClearColor(0, 0, 0.2f, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
-
         stage.act();
         stage.draw();
     }
