+++ core/src/com/mygdx/game/screen/GameScreen.java
@@ -6,14 +6,18 @@ import com.badlogic.gdx.InputMultiplexer;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.OrthographicCamera;
+import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
 import com.badlogic.gdx.scenes.scene2d.Stage;
+import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ActorGestureListener;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
+import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 import com.esotericsoftware.kryonet.Connection;
@@ -21,6 +25,8 @@ import com.esotericsoftware.kryonet.Listener;
 import com.mygdx.game.Carcassonne;
 import com.mygdx.game.GameBoard;
 import com.mygdx.game.Player;
+import com.mygdx.game.Position;
+import com.mygdx.game.tile.Side;
 import com.mygdx.game.network.GameClient;
 import com.mygdx.game.network.NetworkHelper;
 import com.mygdx.game.network.TestOutput;
@@ -57,10 +63,9 @@ public class GameScreen implements Screen {
         this.gameClient = gameClient;
 
         TextButton placeMeeple = new TextButton("place Meeple",  Carcassonne.skin, "default");
-        placeMeeple.setWidth(Gdx.graphics.getWidth()/4);
-        placeMeeple.setHeight(Gdx.graphics.getHeight()/8);
+        placeMeeple.setWidth(Gdx.graphics.getWidth()/4f);
+        placeMeeple.setHeight(Gdx.graphics.getHeight()/8f);
         placeMeeple.getLabel().setFontScale(0.8f);
-        //placeMeeple.setPosition(Gdx.graphics.getWidth()/2-placeMeeple.getWidth()/2, Gdx.graphics.getHeight()*6/9);
         placeMeeple.setPosition(10, 0);
 
         placeMeeple.addListener(new ClickListener() {
@@ -71,24 +76,16 @@ public class GameScreen implements Screen {
                 game.setScreen(new ChosenMeeplePlacementScreen(GameScreen.this,game, gameBoard));
             }
         });
-        //ChosenMeeplePlacementScreen cmpa = new ChosenMeeplePlacementScreen();
         stageUI.addActor(placeMeeple);
-/*
-            @Override
-            public void touchUp(InputEvent event, float x, float y, int pointer, int button) {
-
-                //TODO: neuer Screen öffnet noch nicht
-
-                ChosenMeeplePlacementScreen cmpa = new ChosenMeeplePlacementScreen();
-                stageUI.addActor(cmpa);
-
-                //TODO: SetPosition richtig?
-                cmpa.setPosition(Gdx.graphics.getWidth()-placeMeeple.getWidth(), Gdx.graphics.getHeight());
-            }
-        });*/
-
-        //stageUI.addActor(placeMeeple);
         Gdx.input.setInputProcessor(stage);
+
+        if (NetworkHelper.getGameManager() != null) {
+            NetworkHelper.getGameManager().addListener(new Listener(){
+                public void received (Connection connection, Object object) {
+                    receive(connection,object);
+                }
+            });
+        }
 
         stage.addListener(new InputListener() {
             @Override
@@ -192,4 +189,42 @@ public class GameScreen implements Screen {
         stageUI.dispose();
     }
 
+    public void receive(Connection connection, Object object){
+        //do here what should happen if you get a message of type ...
+        //send message with "Networkhelper.getGameManager.sentToAll(message)
+        //before register the class in the Network class
+        if (object instanceof TestOutput) {
+            //do something
+        }
+    }
+    public static void placeMeeple(GameBoard gameboard, Side side){
+
+        Texture meepleTexture= new Texture(Gdx.files.internal("redmeeple.png"));
+        Position pos = gameboard.getPreviousTile().getPosition();
+        ImageButton meepleImg = new ImageButton(
+                new TextureRegionDrawable(new TextureRegion(meepleTexture)),
+                new TextureRegionDrawable(new TextureRegion(meepleTexture)));
+        meepleImg.setSize(Gdx.graphics.getWidth()/18f,Gdx.graphics.getHeight()/18f);
+        float x = (pos.getX()*128f)+(128f/2f)-(meepleImg.getWidth()/2f);
+        float y = (pos.getY()*128f)+(128f/2f)-(meepleImg.getHeight()/2f);
+
+        switch (side) {
+            case top:
+                y += 42f;
+                break;
+            case left:
+                x -= 42f;
+                break;
+            case right:
+                x += 42f;
+                break;
+            case bottom:
+                y -= 42f;
+                break;
+        }
+
+        meepleImg.setPosition(x, y);
+        //meepleImg.setPosition(calculatePosition(gameboard.getPreviousTile()).getX(), calculatePosition(gameboard.getPreviousTile()).getY());
+        gameboard.getStageOfBoard().addActor(meepleImg);
+    }
 }
