+++ core/src/com/mygdx/game/screen/MainMenuScreen.java
@@ -17,7 +17,6 @@ import com.badlogic.gdx.utils.viewport.ScreenViewport;
 import com.mygdx.game.Carcassonne;
 import com.mygdx.game.network.screen.ClientSearchScreen;
 import com.mygdx.game.network.screen.ServerRoomScreen;
-import com.mygdx.game.screen.CreatePlayersScreen;
 
 public class MainMenuScreen implements Screen {
     private Game game;
@@ -83,6 +82,13 @@ public class MainMenuScreen implements Screen {
         TextButton settButton = new TextButton("Settings", Carcassonne.skin, "menu");
         settButton.setWidth(Gdx.graphics.getWidth() / 4);
         settButton.setPosition(Gdx.graphics.getWidth() / 2-settButton.getWidth() / 2, Gdx.graphics.getHeight() * 6 / 9 - settButton.getHeight() * 3);
+        settButton.addListener(new ClickListener() {
+            @Override
+            public void clicked(InputEvent event, float x, float y) {
+                Gdx.app.debug("touch", "start touch up");
+                game.setScreen(new SettingScreen(game));
+            }
+        });
         stage.addActor(settButton);
 
         Gdx.input.setInputProcessor(stage);
